% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustBatch.R
\name{adjustBatch}
\alias{adjustBatch}
\title{Adjusting the technical effect of the batch effect in RNA-Seq data}
\usage{
adjustBatch(X, batch)
}
\arguments{
\item{X}{a data matrix with n rows representing samples and m columns representing genes.}

\item{batch}{numeric vector of batch value for sample. Length - number of samples.}
}
\value{
A list of the following components:
newX - matrix of corrected data
compBatchNum - component number with batch effect
}
\description{
Adjusting the technical effect of the batch effect in RNA-Seq data
}
