% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustLibsize.R
\name{adjustLibsize}
\alias{adjustLibsize}
\title{Adjusting the technical effect of the library size in RNA-Seq data}
\usage{
adjustLibsize(X, libsize)
}
\arguments{
\item{X}{a data matrix with n rows representing samples and m columns representing genes.}

\item{libsize}{numeric vector of sample library size values. Length - number of samples.}
}
\value{
A list of the following components:
newX - matrix of corrected data
compLibsizeNum - component number with library size
}
\description{
Adjusting the technical effect of the library size in RNA-Seq data
}
