% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindComponents.R
\name{findComp}
\alias{findComp}
\title{Find the number of component with batch-effect and component with library size}
\usage{
findComp(
  X,
  libsize = 0,
  batch = 0,
  ncomp = 5,
  fun = "logcosh",
  alg.typ = "parallel",
  method = c("R", "C")
)
}
\arguments{
\item{X}{a data matrix with n rows representing samples and m columns representing genes.}

\item{libsize}{numeric vector of sample library size values. Length - number of samples.}

\item{batch}{numeric vector of batch value for sample. Length - number of samples.}

\item{ncomp}{number of components}
}
\value{
A list of the following components:
compBatchNum - component number with batch effect
compLibsizeNum - component number with library size
S - weight matrix
A - metagene matrix
}
\description{
Find the number of component with batch-effect and component with library size
}
